Implement authentication: registration and login
Used bcrypt.js to hash and compare the plain text password before it reaches the database with a Sequelize hook function.
Used JWT to encrypt user data. Token is returned on every sign up and sign in. Valid for 30 min. Client can the store it to local storage and on every GraphQL query and mutation use it as authorization.